+++ db/migrate/20181102135958_remove_kg_recycled_paper_from_collection_point.rb
@@ -1,5 +1,5 @@
 class RemoveKgRecycledPaperFromCollectionPoint < ActiveRecord::Migration[5.2]
   def change
-    remove_column :collection_points, :kg_recycled_paper, :int
+    remove_column :collection_points, :kg_recycled_paper, :float
   end
 end
