+++ db/migrate/20181102181633_add_kg_trash_to_collection_point.rb
@@ -1,5 +1,5 @@
 class AddKgTrashToCollectionPoint < ActiveRecord::Migration[5.2]
   def change
-    add_column :collection_points, :kg_trash, :int
+    add_column :collection_points, :kg_trash, :float
   end
 end
